Changed KafkaStreamsAssignment to be a container class
